+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- **Busy days open up in the Weekly Planning recap.** A day with more than four events
+  shows "+N more"; tap it to see the rest of that day in place.
+
 ### Changed
 
 ### Fixed